tests/requirements.txt
# Integration Test Dependencies

pytest==8.0.0
pytest-xdist==3.5.0
boto3==1.35.90
mysql-connector-python==9.0.0
requests==2.32.3